[one-webapp] remove previous forgeweb grpc startup
Move the code to chain startup.
